backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "python-dotenv>=1.1.1",
    "mysqlclient>=2.2.0",
    "gunicorn>=21.2.0",
    "requests>=2.31.0",
]

